+++ jackson/src/test/scala/com/cerner/beadledom/jackson/JacksonModuleSpec.scala
@@ -37,11 +37,11 @@ class JacksonModuleSpec extends AnyFunSpec with Matchers with MockitoSugar {
         modules.iterator().next() should be(new TestModule().getTypeId)
       }
 
-      it("uses snake casing") {
+      it("uses lower camel casing") {
         val model = new CamelCaseModel
         val str = mapper.writeValueAsString(model)
-        str should include("long_field_name_with_camel_case")
-        str should not include "longFieldNameWithCamelCase"
+        str should include("longFieldNameWithCamelCase")
+        str should not include "long_field_name_with_camel_case"
       }
 
       it("omits null fields when serializing") {